weather_airq/requirements.txt
requests==2.32.3
google-cloud-pubsub==2.34.0
